requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg[binary]==3.1.18
psycopg_pool==3.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
tiktoken==0.5.2
openai>=1.6.1
python-dotenv==1.0.0
requests==2.31.0
langchain==0.0.350
langchain-openai==0.0.2
pgvector==0.2.4
pyyaml==6.0.1
prometheus-client==0.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1


